use size_t for element_size in containers

+++ source/queue.h
@@ -5,7 +5,7 @@ typedef struct
 {
     char* arr;
     int len;
-    int element_size;
+    size_t element_size;
 } Queue;
